% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_gravity.R
\name{ge_gravity}
\alias{ge_gravity}
\title{Solves a general equilibrium one sector Armington-CES trade model.}
\usage{
ge_gravity(exp_id, imp_id, flows, beta, theta = 1, mult = FALSE, data = list())
}
\arguments{
\item{exp_id}{String representation of the exporter/origin country associated with each observation.
This is arbitrary and for organization purposes only, i.e. AUS, Australia}

\item{imp_id}{String representation of the importer/destination country associated with each observation.
This is arbitrary and for organization purposes only, i.e. AUS, Australia}

\item{flows}{Observed trade flows in the data for the year being used as the baseline for the counterfactual.}

\item{beta}{An input reflecting the “partial” change in trade, typically obtained as a coefficient from a
prior gravity estimation.}

\item{theta}{Overall trade elasticity}

\item{mult}{If true, assume that national expenditure is a fixed multiple of national output,
as in Anderson & Yotov (2016).
Otherwise (and by default), handle unbalances in the data by treating the
trade balance as an additive component of national expenditure (see below).}

\item{data}{A list to which we should add the new values. By default, this is an empty list.
Note that this will be converted to a named dataframe on output.}
}
\value{
A dataframe element containing resulting estimations of impacts.
   Specifically, it returns results for general equilibrium changes in trade flows,
   welfare, and real wages as a result of the change in trade frictions.
   If `data` is specified, the results will be added as columns.

   This data will include the following:
   \itemize{
    \item X_eu:        The new level of trade for each pair of countries.
    \item welfare:     The exporter's change in welfare (new/old level of welfare)
    \item real_wage:   The exporter's change in real wage (new/old real wage). \cr
       Note: this is generally different from the change in welfare
       unless either trade is balanced or the “multiplicative” option is chosen.
    \item nom_wage:    The exporter's change in nominal wage (new/old nom wage).
    \item price_index: The exporter's change in price index (new/old price index)
   }
}
\description{
\code{ge_gravity} solves for general equilibrium effects of changes in trade policies
  using a one sector Armington-CES trade model. It uses a simple fixed point
  algorithm that allows for fast computation that makes this program
  ideal for bootstrapping confidence intervals for general equilibrium simulations
  based on prior gravity estimates of FTAs or other similar variables.

  Examples of references that conduct general equilibrium analysis based on FTA
  estimates in this way include Egger, Larch, Staub, & Winkelmann (2011),
  Anderson & Yotov (2016), and Baier, Yotov, & Zylkin (2019). Yotov, Piermartini,
  Monteiro, & Larch (2016) provide a detailed survey and introduction to the topic.
}
\details{
Please see `ge_gracity_rmd("all")` for additional details.
}
\references{
Please see `ge_gracity_rmd("all")` for information on references and sources.
}
\seealso{
`ge_gracity_rmd` allows you to access very explanatory RMD files to augment documentation.
}
