% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TradeData0014}
\alias{TradeData0014}
\title{Trade Data 2000 - 2014}
\format{
A list with 7744 rows and 9 variables:
\describe{
  \item{\code{exporter}}{Country code of exporter country}
  \item{\code{importer}}{Country code of importer country}
  \item{\code{expcode }}{Label encoding of exporter country}
  \item{\code{impcode }}{Label encoding of importer country}
  \item{\code{year    }}{Year of row data}
  \item{\code{trade   }}{Aggregate trade flow between importer and exporter for that year}
  \item{\code{eu_enlargement}}{0-1 Booleanic; whether this pair in the EU}
  \item{\code{other_fta     }}{0-1 Booleanic; other FTA pairwise satisfaction}
  \item{\code{FTA           }}{0-1 Booleanic; eu_enlargement or other_fta}
}
}
\source{
\url{http://www.wiod.org/database/wiots16}

\url{https://sites.nd.edu/jeffrey-bergstrand/database-on-economic-integration-agreements/}

\url{http://www.tomzylkin.com/uploads/4/1/0/4/41048809/help_file.pdf}
}
\usage{
TradeData0014
}
\description{
The data set represents the aggregate trade between 44 countries observed over 2000-2014,
using years 2000, 2005, 2010, and 2014.

Trade and domestic sales data has been aggregated from the WIOD database (see sources).
For its use, consider "An Illustrated User Guide to the World Input-Output Database.

Information on Free Trade Agreements (FTAs) is taken from the NSF-Kellogg database
maintained by Scott Baier and Jeff Bergstrand (see sources).
}
\keyword{datasets}
